(refactor): Implement file destructure - destructure files from home page - destructure menu from homeview
[Starts #177933454]
