+++ .env.example
@@ -1,0 +1,3 @@
+APOLLO_KEY=YOUR_API_KEY
+APOLLO_GRAPH_VARIANT=current
+APOLLO_SCHEMA_REPORTING=true
