+++ client/src/.env.example
@@ -1,0 +1,1 @@
+APOLLO_KEY=YOUR_API_KEY
